Refactor Functions into Package
