+++ cli/cmd/runner.go
@@ -135,4 +135,5 @@ type runnerArgs struct {
 
 	enterpriseInstallerAssignInstallerID string
 	enterpriseInstallerAssignChannelID   string
+	initKotsAppSkipPrompt                bool
 }
